box-layout: Fix floating point truncation when calculating a child's size
The child size is a float, not an int.

https://bugzilla.gnome.org/show_bug.cgi?id=707808

+++ clutter/clutter-box-layout.c
@@ -977,7 +977,7 @@ clutter_box_layout_allocate (ClutterLayoutManager   *layout,
   gint extra;
   gint n_extra_widgets = 0; /* Number of widgets that receive 1 extra px */
   gint x = 0, y = 0, i;
-  gint child_size;
+  gfloat child_size;
 
   count_expand_children (layout, container, &nvis_children, &nexpand_children);
